Created wall texture
Created a wall texture, tested it and encoded it inside the wall.tex file. The first line represent the colors that will be used ,each followed by it's index that will be used. The following data represents the index of the color of the texture. The size of the texture is the screen height and width.